Mostly done with assignment 4
Could still use some additional input validation or bonus features

+++ public/js/mainScript.js
@@ -1,6 +1,18 @@
 const container = document.getElementById("container");
+const addBtn = document.getElementById("addBtn");
+const registerBtn = document.getElementById("registerBtn");
+const loginBtn = document.getElementById("loginBtn");
+const logoutBtn = document.getElementById("logoutBtn");
 
 const getList = async(_msg)=>{
+    const loggedin = await isUserAuthenticated();
+    if(loggedin){
+        registerBtn.remove();
+        loginBtn.remove();
+    }else{
+        addBtn.remove();
+        logoutBtn.remove();
+    }
     try {
         const res = await fetch("/food"); //get
         if(!res.ok)throw new Error("Failed to get list");
@@ -11,27 +23,34 @@ const getList = async(_msg)=>{
         list.forEach(e => {
             //Create list item
             const li = document.createElement("li");
-            //Create delete button for list item
-            const delete_btn = document.createElement("button");
-            delete_btn.innerText = "⌫";
-            delete_btn.onclick = (ms)=>{deleteListItem(e.name);};
-            delete_btn.style.marginRight="10px";
-            delete_btn.style.marginLeft="20px";
-            delete_btn.style.maxHeight="20px";
-            delete_btn.style.maxWidth="20px";
-            delete_btn.style.padding="0px";
-            //Create update button for list item
-            const update_btn = document.createElement("button");
-            update_btn.innerText = "⚙️";
-            update_btn.onclick = (ms)=>{editListItem(e._id);};
-            update_btn.style.maxHeight="20px";
-            update_btn.style.maxWidth="20px";
-            update_btn.style.padding="0px";
+
+            let delete_btn,update_btn;
+            if(loggedin){
+                //Create delete button for list item
+                delete_btn = document.createElement("button");
+                delete_btn.innerText = "⌫";
+                delete_btn.onclick = (ms)=>{deleteListItem(e.name);};
+                delete_btn.style.marginRight="10px";
+                delete_btn.style.marginLeft="20px";
+                delete_btn.style.maxHeight="20px";
+                delete_btn.style.maxWidth="20px";
+                delete_btn.style.padding="0px";
+                //Create update button for list item
+                update_btn = document.createElement("button");
+                update_btn.innerText = "⚙️";
+                update_btn.onclick = (ms)=>{editListItem(e._id);};
+                update_btn.style.maxHeight="20px";
+                update_btn.style.maxWidth="20px";
+                update_btn.style.padding="0px";
+                
+            }
             //Finish list item and append buttons
             li.className="food";
             li.innerHTML = `${e.name} <i>${e.foodType}</i>`;
-            li.appendChild(delete_btn);
-            li.appendChild(update_btn);
+            if(loggedin){
+                li.appendChild(delete_btn);
+                li.appendChild(update_btn);
+            }
             _l.appendChild(li);
         });
         if(_msg!=undefined){
@@ -45,17 +64,29 @@ const getList = async(_msg)=>{
 
 const deleteListItem = async(name)=>{
     try {
+        const _a = await fetch("/userauthenticated");
+        if(_a.status!=201){
+            if(_a.status==200) window.location.href = "/login";
+            throw new Error("Failed to authenticate user");
+        }
         const _q = "/deletefood/name?name="+name;
         const res = await fetch(_q,{method:"delete"});
         if(!res.ok)throw new Error("Failed to delete list item");        
         //refresh list
         getList("Successfully deleted "+name);
-    } catch (e) {console.error("Error: ",e); container.innerHTML="<p style='color:red'>Failed to delete list item</p>";}
+    } catch (e) {console.error("Error: ",e); container.innerHTML=`<p style='color:red'>${e}</p>`;}
+}
+
+const isUserAuthenticated = async()=>{
+    try {
+        const _a = await fetch("/userauthenticated");
+        return (_a.status==201)
+    } catch (e) {return false;}
 }
 
 const editListItem = async(id)=>{
     try{
-        window.location.href = "/additem.html?id="+id;
+        window.location.href = "/addfooditem?id="+id;
     }catch(e){console.error("Error: ",e);}
 }
 
@@ -65,6 +96,18 @@ const getEditItem = async(id)=>{
         const res = await fetch(_q,{method:"get"});
         if(!res.ok)throw new Error("Failed to get item to edit");
     } catch (e) {console.error("Error: ",e); container.innerHTML="<p style='color:red'>Failed to delete list item</p>";}
+}
+
+const getSessionUser = async()=>{
+    const loggedin = await isUserAuthenticated();
+    if(!loggedin)return;
+    try{
+        const res = await fetch("/currentuser");
+        if(res.ok){
+            const _user = await res.json();
+            document.getElementById("currUser").innerText="Logged in as: "+_user;
+        }
+    }catch(e){console.error(e);}
 }
 
 const getRandCatImg = async()=>{
@@ -79,5 +122,6 @@ const getRandCatImg = async()=>{
     }catch(e){console.error("Error: ",e);}
 }
 
+getRandCatImg(); //more important so do this first
 getList();
-getRandCatImg();
+getSessionUser();
